rename file & minor fix
